requirement.txt
chardet==3.0.4
Django==2.1.7
django-csvimport==2.12
django-six==1.0.4
djangorestframework==3.9.2
mysqlclient==1.4.2
Pillow==5.4.1
psycopg2-binary==2.7.6.1
pytz==2018.9
xlwt==1.3.0
